requirements.txt
fastapi
uvicorn
httpx
sentence-transformers
faiss-cpu
numpy
pandas
python-dotenv
